Removed pagent and make possible for windows platforms to use UNIXSocket based pagent if available
